Concat images and audio in single ffmpeg step (#35)
Fixes #34.

This seems to work great. CI running times are halved and when
generating subtitles on a video that has 20 slides and takes multiple
minutes all subtitles remain in sync for the duration of the video
(related to https://github.com/transformrs/trv/pull/25) 🚀.

This is an example of the command that the code generates:

```sh
$ ffmpeg \
    -i _out/audio/1.wav \
    -loop 1 -framerate 1 -t "00:00:04.93" -i _out/image/1.png \
    -i _out/audio/2.wav \
    -loop 1 -framerate 1 -t "00:00:07.48" -i _out/image/2.png \
    -filter_complex \
    "
    [1:v]scale=-1:1080,format=yuv420p[v0]; \
    [3:v]scale=-1:1080,format=yuv420p[v1]; \
    [v0][0:a][v1][2:a] concat=n=2:v=1:a=1 [outv] [outa] \
    " \
    -map "[outv]" -r 30 -map "[outa]" -tune stillimage \
    -c:a opus \
    -strict -2 \
    -shortest \
    -movflags \
    +faststart \
    _out/out.mp4 
```